+++ lib/inc/model_matrix.h
@@ -33,7 +33,7 @@ public:
      * @param[in] column matrix column.
      * @param[in] element matrix element.
      */
-    ModelMatrix(const unsigned int row, const unsigned int column, const float *element);
+    ModelMatrix(const unsigned int row, const unsigned int column, const double *element);
 
     /**
      * @brief Create a new ModelMatrix instance.
@@ -41,7 +41,7 @@ public:
      * @param[in] column matrix column.
      * @param[in] element matrix element.
      */
-    ModelMatrix(const unsigned int row, const unsigned int column, const float **element);
+    ModelMatrix(const unsigned int row, const unsigned int column, const double **element);
 
     /**
      * @brief Create a new ModelMatrix instance.
@@ -49,7 +49,7 @@ public:
      * @param[in] column matrix column.
      * @param[in] element matrix element.
      */
-    ModelMatrix(const unsigned int row, const unsigned int column, const std::vector<float> element);
+    ModelMatrix(const unsigned int row, const unsigned int column, const std::vector<double> element);
 
     /**
      * @brief Destructor
@@ -73,7 +73,7 @@ public:
      * @brief get matrix element
      * @returns matrix element.
      */
-    std::vector<float> element() const;
+    std::vector<double> element() const;
 
     /**
      * @brief get matrix element
@@ -81,7 +81,7 @@ public:
      * @param[in] column matrix column.
      * @returns matrix element.
      */
-    float get(const unsigned int row, const unsigned int column) const;
+    double get(const unsigned int row, const unsigned int column) const;
 
     /**
      * @brief set matrix element
@@ -89,7 +89,7 @@ public:
      * @param[in] column matrix column.
      * @param[in] value element value.
      */
-    void set(const unsigned int row, const unsigned int column, const float value);
+    void set(const unsigned int row, const unsigned int column, const double value);
 
     /**
      * @brief calculate zero matrix
@@ -125,7 +125,7 @@ public:
      * @brief calculate determinant
      * @returns determinant.
      */
-    float determinant();
+    double determinant();
 
     /**
      * @brief calculate inverse matrix
@@ -138,13 +138,13 @@ public:
      * @param[in] sigma DLS sigma.
      * @returns result matrix.
      */
-    ModelMatrix inverse(const float sigma);
+    ModelMatrix inverse(const double sigma);
 
     /**
      * @brief get vector length
      * @returns length.
      */
-    float length() const;
+    double length() const;
 
     /**
      * @brief get normalized vector
@@ -156,7 +156,7 @@ public:
      * @brief calculate vector dot product
      * @returns result value.
      */
-    float dot(const ModelMatrix &rhs);
+    double dot(const ModelMatrix &rhs);
 
     /**
      * @brief calculate vector cross product
@@ -171,28 +171,28 @@ public:
     ModelMatrix cross();
 
     ModelMatrix &operator=(const ModelMatrix &other);
-    ModelMatrix operator+(const float &rhs);
+    ModelMatrix operator+(const double &rhs);
     ModelMatrix operator+(const ModelMatrix &rhs);
-    ModelMatrix operator-(const float &rhs);
+    ModelMatrix operator-(const double &rhs);
     ModelMatrix operator-(const ModelMatrix &rhs);
-    ModelMatrix operator*(const float &rhs);
+    ModelMatrix operator*(const double &rhs);
     ModelMatrix operator*(const ModelMatrix &rhs);
 
-    friend ModelMatrix operator+(const float &lhs, const ModelMatrix &rhs);
-    friend ModelMatrix operator-(const float &lhs, const ModelMatrix &rhs);
-    friend ModelMatrix operator*(const float &lhs, const ModelMatrix &rhs);
+    friend ModelMatrix operator+(const double &lhs, const ModelMatrix &rhs);
+    friend ModelMatrix operator-(const double &lhs, const ModelMatrix &rhs);
+    friend ModelMatrix operator*(const double &lhs, const ModelMatrix &rhs);
 
 private:
     ModelMatrix pseudoInverse();
     ModelMatrix pseudoInverseR();
     ModelMatrix pseudoInverseL();
-    float determinant(std::vector<float> matrix, int order);
-    std::vector<float> matrixInversion(std::vector<float> matrix, int order);
+    double determinant(std::vector<double> matrix, int order);
+    std::vector<double> matrixInversion(std::vector<double> matrix, int order);
 
 private:
     unsigned int row_;
     unsigned int column_;
-    std::vector<float> element_;
+    std::vector<double> element_;
 };
 
 #endif // MODEL_MATRIX_H
